+++ idi/reconstruction/cucorrad.py
@@ -115,7 +115,7 @@ def corrfunction(shape, z, maxq, xcenter=None, ycenter=None):
                 jzero(doutput)
                 jkernel(dqx, dqy, dqz, dvals, doutput)
                 jreduce(doutput)
-            return doutput[0, :qmax].copy_to_host(stream=stream)
+            return doutput[0, :qmax].copy_to_host(stream=stream).astype(np.float64)
 
         yield corr
     finally:
